Crate bubblebabble

bubblebabble

bubblebabble is a Rust implementation of the Bubble Babble binary data encoding.

The spec defines the following test vectors:

assert_eq!(
    bubblebabble::encode(&[]),
    String::from("xexax")
);
assert_eq!(
    bubblebabble::encode(&b"1234567890"[..]),
    String::from("xesef-disof-gytuf-katof-movif-baxux")
);
assert_eq!(
    bubblebabble::encode(&b"Pineapple"[..]),
    String::from("xigak-nyryk-humil-bosek-sonax")
);

bubblebabble supports decoding to a byte vector:

assert_eq!(
    bubblebabble::decode("xexax"),
    Ok(vec![])
);
assert_eq!(
    bubblebabble::decode("xesef-disof-gytuf-katof-movif-baxux"),
    Ok(Vec::from(&b"1234567890"[..]))
);
assert_eq!(
    bubblebabble::decode("xigak-nyryk-humil-bosek-sonax"),
    Ok(Vec::from(&b"Pineapple"[..]))
);

License

bubblebabble is licensed under the MIT License (c) Ryan Lopopolo.

bubblebabble is derived from bubble-babble-ts @ v1.0.1. bubble-babble-ts is licensed under the MIT License Copyright (c) 2018 Jonathan M. Wilbur <jonathan@wilbur.space>.

Enums

DecodeError
Decoding errors from bubblebabble::decode.

Functions

decode
Decode a string slice to a vector of bytes.

encode
Encode a vector of bytes with the Bubble Babble encoding to a String.